+++ CHANGELOG.md
@@ -12,6 +12,7 @@
  - Modified effect equations can now use a second parameter, "i", representing total # of iterations
  - Add Plot effect, mainly intended to be used for testing equations, not really for final effects
  - Fix particle offset x/y/z being an integer
+ - ShieldEffect radius changed to floating point
 
 # 5.7